LayoutTests/storage/open-database-creation-callback-isolated-world-expected.txt

ALERT: undefined
This test tests that the openDatabase() creation callback is called in the right world.
